orchestrator/pyproject.toml
[tool.poetry]
name = "ai-investment-orchestrator"
version = "0.1.0"
description = "Autonomous AI investment portfolio manager using local LLM via llama-swap"
authors = ["AI Investment Simulator"]
readme = "README.md"
packages = [{include = "src"}, {include = "dashboard"}]

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.27"
openai = "^1.50"
yfinance = "^0.2"
pandas = "^2.2"
ta = "^0.11"
pydantic = "^2.9"
pydantic-settings = "^2.5"
apscheduler = "^3.10"
structlog = "^24.4"
feedparser = "^6.0"
pyyaml = "^6.0"
streamlit = "^1.40"
plotly = "^5.24"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
pytest-asyncio = "^0.24"
pytest-httpx = "^0.32"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
